Initial check-in
Getting this in for more eyes on it. Still need to implement the refresh of the dashboard  (line 115 in project-management.js). Also, there's an issue now with the use of the xlsx library: any update changes all the dates in our excel file. When this happens, replace tasks.xlsx with 'tasks - master.xlsx'.