Update .gitignore (#110)
* Update .gitignore

Adding Gemfile.lock to the list of files to ignore. The lock file is not needed on GitHub, and it gets generated when you run bundle jekyll server/build locally. Removing it ensures that we have the same dependencies that GitHub uses.

* Delete Gemfile.lock

Not needed in source control. We want to use the GitHub pages gem.  See https://github.com/radar/guides/issues/14