Update datasource.html
This patch is temporary and will be superseded when we update the reference pages. 
This patch updates DataSourceInterfaces.ts  - adding a note to explain that refreshAsync() should not be used to automatically refresh the viz, and that doing so on Tableau Online might result in the removal of the extension.