Doc updates for 2018.3 and release (#183)
* Update readme.md (#152)

Fixed typo of the word 'extension' in third sentence.

* Update trex_configure.md (#156)

Fixed JavaScript typo found by @jakkyn

* Update trex_getstarted.md (#157)

Fixed error found by @jakkyn - In the Choose dialog box, the Choose button is now My Extensions.

* Update trex_overview.md (#160)

Added the "what you can do list" to the overview page

* Doc update to add design guidelines, tableau UI links, and localhost info (#165)

Adding the doc updates for TC18 (Design Guidelines, Tableau UI link).

* Update trex_security.md

adding a comma to help fix a weird sentence.

* Update datasource.html (#170)

* Update datasource.html

This patch is temporary and will be superseded when we update the reference pages. 
This patch updates DataSourceInterfaces.ts  - adding a note to explain that refreshAsync() should not be used to automatically refresh the viz, and that doing so on Tableau Online might result in the removal of the extension.

* Update datasource.html

refined note with John's suggestion (blocked vs. removal)

* Doc updates for 2018.3  (#178)

* updated docs for debugging, error handling, getting started, creating first extension

* Updated debugging gifs, minor doc updates

* Image used for the Extensions API Reference pages

* Fixed typos

* Fixing broken link

* Fixed broken link to error codes topic

* fixed typo

* More tech review updates and CX editorial fixes

* Added Keshia's allow/deny flowchart; more CX edits

* Update trex_publish.md

Adding a note that hosting extensions on Tableau Server is not recommended.

* Update trex_debugging.md

* Update trex_publish.md

Refining the message that the web server needs to be on a different computer than the one running Tableau Server.

* Update trex_publish.md

Removed "the" from "the Tableau Server or Tableau Online."

* Update trex_publish.md

Changed the wording about hosting extensions on Tableau Server back to what it was before the last commit (as not recommended).