Release 1.1.0 (#200)
* Update readme.md (#152)

Fixed typo of the word 'extension' in third sentence.

* Update trex_configure.md (#156)

Fixed JavaScript typo found by @jakkyn

* Update trex_getstarted.md (#157)

Fixed error found by @jakkyn - In the Choose dialog box, the Choose button is now My Extensions.

* Update trex_overview.md (#160)

Added the "what you can do list" to the overview page

* Doc update to add design guidelines, tableau UI links, and localhost info (#165)

Adding the doc updates for TC18 (Design Guidelines, Tableau UI link).

* Update trex_security.md

adding a comma to help fix a weird sentence.

* Update datasource.html (#170)

* Update datasource.html

This patch is temporary and will be superseded when we update the reference pages. 
This patch updates DataSourceInterfaces.ts  - adding a note to explain that refreshAsync() should not be used to automatically refresh the viz, and that doing so on Tableau Online might result in the removal of the extension.

* Update datasource.html

refined note with John's suggestion (blocked vs. removal)

* Doc updates for 2018.3  (#178)

* updated docs for debugging, error handling, getting started, creating first extension

* Updated debugging gifs, minor doc updates

* Image used for the Extensions API Reference pages

* Fixed typos

* Fixing broken link

* Fixed broken link to error codes topic

* fixed typo

* More tech review updates and CX editorial fixes

* Added Keshia's allow/deny flowchart; more CX edits

* Update trex_publish.md

Adding a note that hosting extensions on Tableau Server is not recommended.

* Update trex_debugging.md

* Update trex_publish.md

Refining the message that the web server needs to be on a different computer than the one running Tableau Server.

* Update trex_publish.md

Removed "the" from "the Tableau Server or Tableau Online."

* Update trex_publish.md

Changed the wording about hosting extensions on Tableau Server back to what it was before the last commit (as not recommended).

* Doc updates for 2019.1 and v1.2 library; show-hide, chromium, API syn… (#195)

* Doc updates for 2019.1 and v1.2 library; show-hide, chromium, API sync fix, best practices

* removed best practices, changed to 1.1, modified visbility-error workaround

* correcting version number to 1.1

* fixing library 1.1 references; drop-down menu fix

* Update ux_controls_ui_patterns.md (#199)

Adding a heading to a column in a table. The Markdown looked fine, but Jekyll didn't recognize it as a table without a heading to the column.

* Fixing missing .gif file

Added the .gif file with the correct name. The previous file name had a `(`` instead of an `_`.

* Delete 3-Small)on_Dark_Background.gif

Had a typo in the file name. Added the file with the correct name.

* Release 1.1.0