Adding PDF Viewer sample
Adding PDF Viewer sample to the Typescript samples.  This PDF Viewer sample allows pdfs to be viewed using a dashboard network extension.  It makes use of extension-apis around settings, dialogs, and parameters.

The sample adds support for testing using jest and adding React components from tableau-ui.  Adding support for tableau-ui requires React versions to match so had to move this project's react from 17.0.2 to 16.8.6.  React-types didn't have a 16.8.6 version so went with the previous version of 16.8.5.